Updates Stripe iOS SDK to 24.23.0
As a result, this fixes the following issue:

“Problems collecting bank account payment method
I was able to test the credit card payment method collection.
But if I call the SDK with collectPaymentMethod("BankAccount", {}) it ends up showing me the credit card payment method collection as well.”

And this also adapts to the latest performCheckout API changes in the native SDK, which throws errors instead of using a failure case on CheckoutResult, and also supports a cancellation result.